+++ src/Runtime/Error.hs
@@ -1,10 +1,12 @@
 module Runtime.Error (EvalError (..), displayEvalErr) where
 
+import Data.Text (Text, unpack)
+
 data EvalError = EvalError
   { errorLine :: Int,
-    errorMessage :: String
+    errorMessage :: Text
   }
   deriving stock (Show, Eq)
 
 displayEvalErr :: EvalError -> String
-displayEvalErr (EvalError line msg) = msg <> "\n[line " <> show line <> "]"
+displayEvalErr (EvalError line msg) = unpack msg <> "\n[line " <> show line <> "]"